Creating a separate file for IUserService interface
