+++ Services/UserService.cs
@@ -6,16 +6,6 @@ using RedBrain.Authentication.Helpers;
 
 namespace RedBrain.Authentication.Services
 {
-    public interface IUserService
-    {
-        User Authenticate(string username, string password, string tenant);
-        IEnumerable<User> GetAll(string tenant);
-        User GetById(int id);
-        User Create(User user, string password);
-        void Update(User user, string password = null);
-        void Delete(int id);
-    }
-
     public class UserService : IUserService
     {
         private DataContext _context;
